test: split and refactor jenkins module tests(agents, boostrap, plugins)
